refactored for cleaner separation of functions
